+++ custom_components/ford_triplog/history.py
@@ -3,7 +3,7 @@ Ford Triplog
 
 History and statistics.
 
-Version: 1.6.0
+Version: 1.6.3
 """
 
 from __future__ import annotations